+++ src/Ryujinx/Headless/HeadlessRyujinx.Init.cs
@@ -224,6 +224,7 @@ namespace Ryujinx.Headless
                         !options.DisablePTC,
                         ITickSource.RealityTickScalar,
                         options.EnableInternetAccess,
+                        options.EnableSwitchNet,
                         !options.DisableFsIntegrityChecks ? IntegrityCheckLevel.ErrorOnInvalid : IntegrityCheckLevel.None,
                         options.FsGlobalAccessLogMode,
                         options.SystemTimeOffset,