+++ src/Ryujinx/Systems/Configuration/ConfigurationState.cs
@@ -79,6 +79,7 @@ namespace Ryujinx.Ava.Systems.Configuration
                 EnableLowPowerPtc = System.EnableLowPowerPtc,
                 TickScalar = System.TickScalar,
                 EnableInternetAccess = System.EnableInternetAccess,
+                EnableSwitchNet = System.EnableSwitchNet,
                 EnableFsIntegrityChecks = System.EnableFsIntegrityChecks,
                 FsGlobalAccessLogMode = System.FsGlobalAccessLogMode,
                 AudioBackend = System.AudioBackend,
@@ -216,6 +217,7 @@ namespace Ryujinx.Ava.Systems.Configuration
             Graphics.ScalingFilterLevel.Value = 80;
             System.EnablePtc.Value = true;
             System.EnableInternetAccess.Value = false;
+            System.EnableSwitchNet.Value = false;
             System.EnableFsIntegrityChecks.Value = true;
             System.FsGlobalAccessLogMode.Value = 0;
             System.AudioBackend.Value = AudioBackend.SDL3;